+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 8.0.24
+
+- Added option to add primary stats (Agi/Int/Str) to the score calculations. In Azerite item +1 stat = +1 score in calculations.
+- This adds 2 new lines for translators at ([Curseforge Localization tool](https://wow.curseforge.com/projects/azeritepowerweights/localization))
+
 ### 8.0.23
 
 - Updating Russian translations